Update header.html to return name to FSA Design System
The header title was changed to FPAC Design System when FSA was merged into FPAC, but the URL at which it lived, the component names, etc. all still referenced FSA. Now that we are coming out with an FPAC Design System we are reverting the name here to avoid confusion.